Added functionality to resize images over 800 pixels wide.

+++ app/classes/AppHelper.php
@@ -5,15 +5,23 @@ class AppHelper {
   {
     $result['result'] = false;
 
-    $file_obj = Image::make($file->getRealPath());
+    $file_path = $file->getRealPath();
+    $file_obj = Image::make($file_path);
     $S3 = App::make('aws')->get('s3');
     $image_path = "images/photos/{$user->id}/1/";
     $new_file_name = str_replace('/', '', Hash::make($file->getClientOriginalName()));
     $thumb_file_name = $new_file_name . '_thumb.' . $file->getClientOriginalExtension();
 
+    // Resize image if larger than 800 pixels wide
+    if($file_obj->width > 800)
+    {
+      $file_path = public_path() . '/' . $new_file_name . $file->getClientOriginalExtension();
+      $file_obj->resize(800, null, true)->save($file_path);
+    }
+
+    // Create thumbnail version
     if($options['thumb'] == true)
     {
-      // Create thumbnail version
       $thumb_width = (array_key_exists('width', $options['thumb_size']) && is_int($options['thumb_size']['width'])) ? $options['thumb_size']['width'] : 64;
       $thumb_height = (array_key_exists('height', $options['thumb_size']) && is_int($options['thumb_size']['height'])) ? $options['thumb_size']['height'] : 64;
 
@@ -23,7 +31,7 @@ class AppHelper {
     // Upload file to AWS S3
     try {
       // Save image to AWS S3
-      $S3->upload(Config::get('constant.AWS.bucket'), $image_path . $new_file_name . '.' . $file->getClientOriginalExtension(), fopen($file->getRealPath(), 'r'), 'public-read');
+      $S3->upload(Config::get('constant.AWS.bucket'), $image_path . $new_file_name . '.' . $file->getClientOriginalExtension(), fopen($file_path, 'r'), 'public-read');
 
       if($options['thumb'] == true)
       {
@@ -43,11 +51,16 @@ class AppHelper {
 
     }
 
-
-    if($options['thumb'] == true)
+    // Destroy temp files if exists after upload
+    if($options['thumb'] == true && file_exists(public_path() . '/' . $thumb_file_name))
     {
       // Destroy local copy of generated thumbnl
       unlink(public_path() . '/' . $thumb_file_name);
+    }
+
+    if($file_obj->width > 800 && file_exists($file_path))
+    {
+      unlink($file_path);
     }
 
     // Destroy file object
